Add a type index per generation.
This is sort of one of the core ideas here, I want the entire dex to be
browsable either as an overview of ALL THE DATA, or one generation at a
time, hence all the generation stuff in porydex.

I debated having the selected gen just be a session thing, stored in a
cookie, but I think it'll make more sense to present it as separate
dexes per generation — if people specifically bookmark a Pokémon's Gen I
dex entry, or link it to their friends, they'll probably want it to
specifically lead to that gen's page.