Fix request.generation attribute, sort of.
It was getting called before traversal for whatever reason, and
therefore getting set to None no matter what.

As indicated in the comments, it still shouldn't be done this way; I
want to change traversal to use separate resources per generation, so
all the index classes would keep the generation on themselves and the db
objects would get wrapped in some container class that holds the
generation and the object.