Change request.session to request.db.
The Pyramid Debug Toolbar does some kind of shenanigans with request.session.
I'm not sure what it expects it to be, but it sure doesn't seem to expect an
SQLA session; it seemed to be trying to treat it like a dict full of string
keys?  And anyway I was getting errors.